Hi Jonny,
here is the latest update. I have change the GFMRPlanner so that we pass the GFExistentialExpression. For that, I add a new class MyGFParser to parse the string and construct the original GFExistentialExpression. I no longer use GFSerializer.

Another difference is that: I add the method prefixString() in GFExpression that return the String of the GFExpression in Prefix/Polish form.
In this form we can parse the String more easily. That's why MyGFParser is much shorter.
Let me know when we can meet to discuss our next step.
-tony